[mlir] Fix -Wsign-compare in ResolveShapedTypeResultDims.cpp (NFC)
/llvm-project/mlir/lib/Dialect/MemRef/Transforms/ResolveShapedTypeResultDims.cpp:98:19: error: comparison of integers of different signs: 'value_type' (aka 'long long') and 'size_t' (aka 'unsigned long') [-Werror,-Wsign-compare]
    if (*dimIndex >= reifiedResultShapes[resultNumber].size())
        ~~~~~~~~~ ^  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ mlir/lib/Dialect/MemRef/Transforms/ResolveShapedTypeResultDims.cpp
@@ -95,7 +95,7 @@ struct DimOfReifyRankedShapedTypeOpInterface : public OpRewritePattern<OpTy> {
       return failure();
     unsigned resultNumber = dimValue.getResultNumber();
     // Do not apply pattern if the IR is invalid (dim out of bounds).
-    if (*dimIndex >= reifiedResultShapes[resultNumber].size())
+    if ((size_t)(*dimIndex) >= reifiedResultShapes[resultNumber].size())
       return rewriter.notifyMatchFailure(dimOp, "dimension is out of bounds");
     Value replacement = getValueOrCreateConstantIndexOp(
         rewriter, dimOp.getLoc(), reifiedResultShapes[resultNumber][*dimIndex]);
